port to import_flickr.py
